#25 Extracted Name Utility functions in Name Core module
