+++ apps/core/lib/name.ex
@@ -11,9 +11,12 @@ defmodule Core.Name do
     NameRevokeTx,
     NameTransferTx,
     NameUpdateTx,
-    GenericSignedTx
+    GenericSignedTx,
+    CommitmentId,
+    NameEntry
   }
 
+  alias AeternityNode.Api.NameService
   alias Core.Client
   alias Utils.Account, as: AccountUtil
   alias Utils.Transaction
@@ -26,7 +29,7 @@ defmodule Core.Name do
   ## Examples
       iex> name = "a123.test"
       iex> name_salt = 7
-      iex> commitment_id = Utils.Name.commitment_id(client, name, name_salt)
+      iex> commitment_id = Core.Name.commitment_id(client, name, name_salt)
       iex> Core.Name.preclaim(client, commitment_id, [fee:  1_000_000_000_000_000])
       {:ok,
        %{
@@ -292,6 +295,37 @@ defmodule Core.Name do
       {:ok, Map.from_struct(tx)}
     else
       error -> {:error, "#{__MODULE__}: Unsuccessful post of NameRevokeTx: #{inspect(error)}"}
+    end
+  end
+
+  @doc """
+  Creates name commitment hash .
+
+  ## Examples
+      iex> Core.Name.commitment_id(client, "a123.test", 7)
+      "cm_2rxmhXWBzjsXTsLfxYK5dqGxKzcJphSijJ3TVHe23mVWYJZhTY"
+  """
+  @spec commitment_id(Client.t(), String.t(), integer()) :: String.t() | {:error, String.t()}
+  def commitment_id(%Client{connection: connection}, name, name_salt)
+      when is_binary(name) and is_integer(name_salt) do
+    case NameService.get_commitment_id(connection, name, name_salt) do
+      {:ok, %CommitmentId{commitment_id: commitment_id}} -> commitment_id
+      error -> error
+    end
+  end
+
+  @doc """
+  Lookups the information about the given name.
+
+  ## Examples
+      iex> Core.Name.get_name_id_by_name(client, "a123.test")
+      {:ok, %AeternityNode.Model.Error{reason: "Name revoked"}}
+  """
+  @spec get_name_id_by_name(Client.t(), String.t()) :: String.t() | {:error, String.t()}
+  def get_name_id_by_name(%Client{connection: connection}, name) when is_binary(name) do
+    case NameService.get_name_entry_by_name(connection, name) do
+      {:ok, %NameEntry{id: id}} -> id
+      error -> error
     end
   end
