+++ apps/utils/mix.exs
@@ -33,7 +33,6 @@ defmodule Utils.Mixfile do
   defp deps do
     [
       {:enacl, github: "aeternity/enacl", ref: "26180f42c0b3a450905d2efd8bc7fd5fd9cece75"},
-      {:core, in_umbrella: true},
       {:erl_base58, "~> 0.0.1"}
     ]
   end
